+++ os/src/trap/mod.rs
@@ -51,7 +51,7 @@ fn set_kernel_trap_entry() {
 /// set trap entry for traps happen in user mode
 fn set_user_trap_entry() {
     unsafe {
-        stvec::write(TRAMPOLINE as usize, TrapMode::Direct);
+        stvec::write(TRAMPOLINE, TrapMode::Direct);
     }
 }
